Upgraded version of d3node-output.
Was running into some issues with svg2png so I upgraded the version of
d3node-output, which fixed the issue (mostly through switching that
dependency to puppeteer I believe).  Also installed yarn and ran 'yarn
install' so as to update 'yarn.lock' correspondingly - this appears to
have created sizeable changes in that lock file.